+++ analytics.js
@@ -907,15 +907,15 @@ exports.parse = function(url){
   a.href = url;
   return {
     href: a.href,
-    host: a.host || location.host,
-    port: ('0' === a.port || '' === a.port) ? port(a.protocol) : a.port,
+    host: a.host,
+    port: a.port,
     hash: a.hash,
-    hostname: a.hostname || location.hostname,
-    pathname: a.pathname.charAt(0) != '/' ? '/' + a.pathname : a.pathname,
-    protocol: !a.protocol || ':' == a.protocol ? location.protocol : a.protocol,
+    hostname: a.hostname,
+    pathname: a.pathname,
+    protocol: a.protocol,
     search: a.search,
     query: a.search.slice(1)
-  };
+  }
 };
 
 /**
@@ -927,7 +927,9 @@ exports.parse = function(url){
  */
 
 exports.isAbsolute = function(url){
-  return 0 == url.indexOf('//') || !!~url.indexOf('://');
+  if (0 == url.indexOf('//')) return true;
+  if (~url.indexOf('://')) return true;
+  return false;
 };
 
 /**
@@ -939,7 +941,7 @@ exports.isAbsolute = function(url){
  */
 
 exports.isRelative = function(url){
-  return !exports.isAbsolute(url);
+  return ! exports.isAbsolute(url);
 };
 
 /**
@@ -952,29 +954,10 @@ exports.isRelative = function(url){
 
 exports.isCrossDomain = function(url){
   url = exports.parse(url);
-  return url.hostname !== location.hostname
-    || url.port !== location.port
-    || url.protocol !== location.protocol;
+  return url.hostname != location.hostname
+    || url.port != location.port
+    || url.protocol != location.protocol;
 };
-
-/**
- * Return default port for `protocol`.
- *
- * @param  {String} protocol
- * @return {String}
- * @api private
- */
-function port (protocol){
-  switch (protocol) {
-    case 'http:':
-      return 80;
-    case 'https:':
-      return 443;
-    default:
-      return location.port;
-  }
-}
-
 });
 require.register("component-bind/index.js", function(exports, require, module){
 /**
@@ -1022,13 +1005,8 @@ module.exports = function (obj) {
 });
 require.register("ianstormtaylor-bind/index.js", function(exports, require, module){
 
-try {
-  var bind = require('bind');
-} catch (e) {
-  var bind = require('bind-component');
-}
-
-var bindAll = require('bind-all');
+var bind = require('bind')
+  , bindAll = require('bind-all');
 
 
 /**
@@ -1104,7 +1082,6 @@ else if (typeof window == 'undefined' || window.ActiveXObject || !window.postMes
 
 });
 require.register("ianstormtaylor-callback/index.js", function(exports, require, module){
-
 var next = require('next-tick');
 
 
@@ -2389,6 +2366,7 @@ var integrations = [
   'uservoice',
   'vero',
   'visual-website-optimizer',
+  'webengage',
   'woopra',
   'yandex-metrica'
 ];
@@ -4197,6 +4175,7 @@ var GA = exports.Integration = integration('Google Analytics')
   .option('doubleClick', false)
   .option('enhancedLinkAttribution', false)
   .option('ignoreReferrer', null)
+  .option('includeSearch', false)
   .option('siteSpeedSampleRate', null)
   .option('trackingId', '')
   .option('trackNamedPages', true)
@@ -4288,7 +4267,7 @@ GA.prototype.page = function (category, name, properties, options) {
   if (name && category) name = category + ' ' + name;
 
   window.ga('send', 'pageview', {
-    page: properties.path,
+    page: path(properties, this.options),
     title: name || properties.title,
     url: properties.url
   });
@@ -4415,7 +4394,7 @@ GA.prototype.pageClassic = function (category, name, properties, options) {
   options = options || {};
   this._category = category; // store for later
 
-  push('_trackPageview', properties.path);
+  push('_trackPageview', path(properties, this.options));
 
   // categorized pages
   if (category && this.options.trackCategorizedPages) {
@@ -4451,6 +4430,21 @@ GA.prototype.trackClassic = function (event, properties, options) {
 
   push('_trackEvent', category, event, label, value, noninteraction);
 };
+
+
+/**
+ * Return the path based on `properties` and `options`.
+ *
+ * @param {Object} properties
+ * @param {Object} options
+ */
+
+function path (properties, options) {
+  if (!properties) return;
+  var str = properties.path;
+  if (options.includeSearch && properties.search) str += properties.search;
+  return str;
+}
 
 
 /**
@@ -4703,7 +4697,9 @@ var traitAliases = {
  */
 
 Heap.prototype.identify = function (id, traits, options) {
+  traits = traits || {};
   traits = alias(traits, traitAliases);
+  if (!traits.handle && id) traits.handle = id;
   window.heap.identify(traits);
 };
 
@@ -4721,6 +4717,7 @@ Heap.prototype.identify = function (id, traits, options) {
 Heap.prototype.track = function (event, properties, options) {
   window.heap.track(event, properties);
 };
+
 });
 require.register("segmentio-analytics.js-integrations/lib/hittail.js", function(exports, require, module){
 
@@ -8295,6 +8292,69 @@ function variation (id) {
   return variationId ? experiment.comb_n[variationId] : null;
 }
 });
+require.register("segmentio-analytics.js-integrations/lib/webengage.js", function(exports, require, module){
+
+var integration = require('integration');
+var load = require('load-script');
+
+/**
+ * Expose plugin
+ */
+
+module.exports = exports = function(analytics){
+  analytics.addIntegration(WebEngage);
+};
+
+/**
+ * Expose `WebEngage` integration
+ */
+
+var WebEngage = exports.Integration = integration('WebEngage')
+  .assumesPageview()
+  .readyOnLoad()
+  .global('_weq')
+  .global('webengage')
+  .option('widgetVersion', '4.0')
+  .option('licenseCode', '');
+
+/**
+ * Initialize.
+ *
+ * @param {Object} page
+ */
+
+WebEngage.prototype.initialize = function(page){
+  var _weq = window._weq = window._weq || {};
+  _weq['webengage.licenseCode'] = this.options.licenseCode;
+  _weq['webengage.widgetVersion'] = this.options.widgetVersion;
+  this.load();
+};
+
+/**
+ * Loaded?
+ *
+ * @return {Boolean}
+ */
+
+WebEngage.prototype.loaded = function(){
+  return !! window.webengage;
+};
+
+/**
+ * Load
+ *
+ * @param {Function} fn
+ */
+
+WebEngage.prototype.load = function(fn){
+  var path = '/js/widget/webengage-min-v-4.0.js';
+  load({
+    https: 'https://ssl.widgets.webengage.com' + path,
+    http: 'http://cdn.widgets.webengage.com' + path
+  }, fn);
+};
+
+});
 require.register("segmentio-analytics.js-integrations/lib/woopra.js", function(exports, require, module){
 
 var each = require('each');
@@ -9634,7 +9694,7 @@ var analytics = module.exports = exports = new Analytics();
  * Expose `VERSION`.
  */
 
-exports.VERSION = '1.1.7';
+exports.VERSION = '1.1.9';
 
 
 /**
@@ -9994,7 +10054,8 @@ Analytics.prototype.page = function (category, name, properties, options, fn) {
     path: canonicalPath(),
     referrer: document.referrer,
     title: document.title,
-    url: location.href
+    url: location.href,
+    search: location.search
   };
 
   if (name) defs.name = name;
@@ -10485,7 +10546,7 @@ Entity.prototype._setTraits = function (traits) {
 
 
 /**
- * Idenfity the entity with an `id` and `traits`. If we it's the same entity,
+ * Identify the entity with an `id` and `traits`. If we it's the same entity,
  * extend the existing `traits` instead of overwriting.
  *
  * @param {String} id
@@ -10965,6 +11026,7 @@ require.alias("segmentio-analytics.js-integrations/lib/userfox.js", "analytics/d
 require.alias("segmentio-analytics.js-integrations/lib/uservoice.js", "analytics/deps/integrations/lib/uservoice.js");
 require.alias("segmentio-analytics.js-integrations/lib/vero.js", "analytics/deps/integrations/lib/vero.js");
 require.alias("segmentio-analytics.js-integrations/lib/visual-website-optimizer.js", "analytics/deps/integrations/lib/visual-website-optimizer.js");
+require.alias("segmentio-analytics.js-integrations/lib/webengage.js", "analytics/deps/integrations/lib/webengage.js");
 require.alias("segmentio-analytics.js-integrations/lib/woopra.js", "analytics/deps/integrations/lib/woopra.js");
 require.alias("segmentio-analytics.js-integrations/lib/yandex-metrica.js", "analytics/deps/integrations/lib/yandex-metrica.js");
 require.alias("segmentio-analytics.js-integrations/index.js", "integrations/index.js");