+++ History.md
@@ -1,3 +1,14 @@
+1.1.9 - December 11, 2013
+-------------------------
+* upgrade `analytics.js-integrations` to `0.2.16`
+* add `search` to page property defaults
+
+1.1.8 - December 11, 2013
+------------------------
+* upgrade `analytics.js-integrations` to `0.2.15`
+* add [WebEngage](http://webengage.com)
+* heap: fallback to user id as handle
+
 1.1.7 - December 4, 2013
 ------------------------
 * upgrade `analytics.js-integrations` to `0.2.13`
@@ -11,7 +22,7 @@
 
 1.1.5 - November 26, 2013
 -------------------------
-* update `analytics.js-integration` to `0.1.5` 
+* update `analytics.js-integration` to `0.1.5`
 * update `analytics.js-integrations` to `0.2.11`
 
 1.1.4 - November 25, 2013